Crate squashfs_ng

This crate provides Rust bindings for the squashfs-tools-ng library, providing support for SquashFS as an embeddable archive format without the need for kernel support. It also tries to provide a level of safety and abstraction on top of the C library. Cross-platform usability is a secondary goal.

Installation

Currently, the underlying squashfs-tools-ng library must be installed on the system both to build and to use this library. The development headers (/usr/include/sqfs/...) are required to build, and the shared library (/usr/lib/libsquashfs.so) to run. The project’s GitHub page asserts that packages are available in many Linux distributions’ repositories.

Once the dependencies are in place, this should function like most other Rust libraries, and cargo build should suffice to build the library.

Usage

The read and write modules below provide support for reading and writing SquashFS files, respectively. Check them out for further documentation.

Modules

bindings 🔒

read

Facilities for reading SquashFS archives.

write

Facilities for writing SquashFS archives.

Structs

ManagedPointer 🔒

Enums

LibError

Errors raised by the underlying library.

SquashfsError

Errors encountered while reading or writing an archive.

Constants

BLOCK_BUF_SIZE 🔒

LINK_MAX 🔒

LOCK_ERR 🔒

NO_XATTRS 🔒

PAD_TO 🔒

_IMPL_NUM_FromPrimitive_FOR_LibError 🔒

Functions

libc_free 🔒

os_to_string 🔒

rust_dealloc 🔒

sfs_check 🔒

sfs_destroy 🔒

sfs_init 🔒

sfs_init_check_null 🔒

sfs_init_ptr 🔒

unpack_meta_ref 🔒

Type Aliases

BoxedError 🔒

Result

Result type returned by SquashFS library operations.