Enum squashfs_ng::LibError

#[repr(i32)]
pub enum LibError {

    Alloc = -1,
    Io = -2,
    Compressor = -3,
    Internal = -4,
    Corrupted = -5,
    Unsupported = -6,
    Overflow = -7,
    OutOfBounds = -8,
    SuperMagic = -9,
    SuperVersion = -10,
    SuperBlockSize = -11,
    NotDir = -12,
    NoEntry = -13,
    LinkLoop = -14,
    NotFile = -15,
    ArgInvalid = -16,
    Sequence = -17,

}

Errors raised by the underlying library.

This error type reflects all errors raised by the squashfs-tools-ng library. This should always be wrapped in a SquashfsError before being returned from any of the functions in this library.

Variants

Alloc = -1

Io = -2

Compressor = -3

Internal = -4

Corrupted = -5

Unsupported = -6

Overflow = -7

OutOfBounds = -8

SuperMagic = -9

SuperVersion = -10

SuperBlockSize = -11

NotDir = -12

NoEntry = -13

LinkLoop = -14

NotFile = -15

ArgInvalid = -16

Sequence = -17