Struct squashfs_ng::write::Writer

pub struct Writer {

    outfile: ManagedPointer<sqfs_file_t>,
    compressor_config: sqfs_compressor_config_t,
    compressor: ManagedPointer<sqfs_compressor_t>,
    superblock: sqfs_super_t,
    block_writer: ManagedPointer<sqfs_block_writer_t>,
    block_processor: Mutex<ManagedPointer<sqfs_block_processor_t>>,
    frag_table: ManagedPointer<sqfs_frag_table_t>,
    id_table: Mutex<ManagedPointer<sqfs_id_table_t>>,
    xattr_writer: Mutex<ManagedPointer<sqfs_xattr_writer_t>>,
    inode_writer: ManagedPointer<sqfs_meta_writer_t>,
    dirent_writer: ManagedPointer<sqfs_meta_writer_t>,
    dir_writer: ManagedPointer<sqfs_dir_writer_t>,
    nodes: Mutex<Vec<RefCell<IntermediateNode>>>,
    finished: RwLock<bool>,

}

A basic SquashFS writer.

This provides a simple interface for writing archives. The user calls open, add to add each node, and finish to finish writing. This is intended for writing archives that are generated by code or otherwise not reflected by files in a file system – if you want to archive a tree of files from disk, TreeProcessor handles directory tracking so that you don’t have to do it yourself.

Each node must be written before its parent, and an error will be raised if this invariant is not maintained – however, this is not detected until finish is called.

let writer = Writer::open("archive.sfs")?;
let mut ids = HashMap::new();
for i in 0..5 {
    let mut content = format!("This is the content of file {}.txt.", i).as_bytes();
    let source = Source::defaults(SourceData::File(Box::new(content)));
    let id = writer.add(source)?;
    ids.insert(OsString::from(format!("{}.txt", i)), id);
}
writer.add(Source::defaults(SourceData::Dir(Box::new(ids.into_iter()))))?;
writer.finish()?;

Fields
outfile: ManagedPointer<sqfs_file_t>compressor_config: sqfs_compressor_config_tcompressor: ManagedPointer<sqfs_compressor_t>superblock: sqfs_super_tblock_writer: ManagedPointer<sqfs_block_writer_t>block_processor: Mutex<ManagedPointer<sqfs_block_processor_t>>frag_table: ManagedPointer<sqfs_frag_table_t>id_table: Mutex<ManagedPointer<sqfs_id_table_t>>xattr_writer: Mutex<ManagedPointer<sqfs_xattr_writer_t>>inode_writer: ManagedPointer<sqfs_meta_writer_t>dirent_writer: ManagedPointer<sqfs_meta_writer_t>dir_writer: ManagedPointer<sqfs_dir_writer_t>nodes: Mutex<Vec<RefCell<IntermediateNode>>>finished: RwLock<bool>
Implementations

impl Writer

pub fn open<T: AsRef<Path>>(path: T) -> Result<Self>

Open a new output file for writing.

If the file exists, it will be overwritten.

fn mode_from_inode(inode: &ManagedPointer<sqfs_inode_generic_t>) -> u16

fn outfile_size(&self) -> u64

pub fn add(&mut self, source: Source) -> Result<u32>

Add the provided Source to the archive.

This writes file data and xattrs to the archive directly, while storing directory tree information to write when finish is called.

The returned value is the inode number of the added Source. If the file is to be added to a directory (that is, almost always), this number needs to be stored so that it can be provided when the directory is added. In the current implementation, inode numbers start at 1 for the first file and count steadily upward, but this behavior may change without warning.

pub fn finish(&mut self) -> Result<()>

Finish writing the archive and flush all contents to disk.

It is an error to call add after this has been run.